% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_quantile_from_prob.R
\name{calc_quantile_from_prob}
\alias{calc_quantile_from_prob}
\title{Calculate quantiles from probabilities}
\usage{
calc_quantile_from_prob(
  prob,
  mu = 0,
  sd = 1,
  n = 1,
  test = "z",
  direction = "greater",
  shade = FALSE,
  digits = 4
)
}
\arguments{
\item{prob}{the probability referring to percentile of a distribution.}

\item{mu}{the mean of the distribution.}

\item{sd}{the standard deviation of the distribution.}

\item{n}{the sample size.  Defaults to 1 for normal distributions.}

\item{test}{takes either "z" or "t" to define the distribution from which the quantile should be pulled.}

\item{direction}{where the probability refers to.  Direction can take any string value, but any string containing the letter "l" will be assumed to mean the "lower" or "left" tail.  Probabilities are assumed to be either tail.}

\item{shade}{a logical for if a shaded density curve should be provided.}

\item{digits}{an integer in for the number of decimal places to which the answer should be rounded.  Note: positive numbers refer to the number of digits after the decimal place, negative numbers refer to the number of digits before the decimal place, and 0 refers to the nearest whole number.}
}
\value{
a statement laying out the values imputed, the subsequent quantile, p-value, and an appropriate conclusion.
}
\description{
Calculates z, t, and non-standardized values for the normal and t distributions from probabilities and the direction the probability refers to.   This function is targeted to help introductory level students and professors to use easily answer questions and learn R, without having to understand for complicated functions like qnorm and qt.
}
\examples{
calc_quantile_from_prob(.9, mu = 20.8, sd = 5.8, direction = "lower", test = "z")
calc_quantile_from_prob(.1, mu = 20.8, sd = 5.8, direction = "lower", test = "z")
}
