% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_t_score.R
\name{calc_t_score}
\alias{calc_t_score}
\title{Calculate t-score and p-values}
\usage{
calc_t_score(
  x,
  mu,
  s,
  n,
  sig_level = NULL,
  p_val = FALSE,
  alternative = 3,
  shade = FALSE,
  digits = 4
)
}
\arguments{
\item{x}{mean or value for which a z-score or probability of observing is to be calculated.}

\item{mu}{mean of the distribution or the hypothesized value.}

\item{s}{standard deviation of the sample}

\item{n}{sample size if x is a mean.  n = 1 by default for the case where x is a single value.}

\item{sig_level}{the significance level for a hypothesis test.}

\item{p_val}{a logical for if a p-value should be provided.}

\item{alternative}{takes 3 values indicating the direction of the alternative hypothesis, or the direction of the probability to be calculated.  The value of this does not matter if p_val is FALSE.  The options for this input are 1 = "greater than", 2 = "less than", and 3 = "not equal to".}

\item{shade}{a logical for if a shaded density curve should be provided.}

\item{digits}{an integer in for the number of decimal places to which the answer should be rounded.  Note: positive numbers refer to the number of digits after the decimal place, negative numbers refer to the number of digits before the decimal place, and 0 refers to the nearest whole number.}
}
\value{
a statement laying out the values imputed, the subsequent t-score, p-value, and an appropriate conclusion (if p_val = TRUE, and sig_level has a value).
}
\description{
Calculates t-score and p-values for hypothesis test and t-score questions.  This function is targeted to help introductory level students and professors to use easily answer questions and learn R, without having to understand for complicated functions like pt and dt.
}
\examples{
calc_t_score(x = 80, mu = 100, s = 50, n = 40, p_val = TRUE, sig_level = 0.05,
alternative = 3)
calc_t_score(x = 10, mu = 5, s = 50, n = 40)

}
