% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade_area.R
\name{shade_area}
\alias{shade_area}
\title{Shades the area of a Normal or T density curve}
\usage{
shade_area(
  start = NULL,
  stop = NULL,
  prob = NULL,
  mu = 0,
  sd = 1,
  n = 1,
  direction = "greater",
  test = "z",
  fill = "orange",
  digits = 4
)
}
\arguments{
\item{start}{the number indicating where the shading starts.  If only tails (one or both) are being shaded, the start should be the value closest to the mean (the function shades from the middle to +/- infinity).  If a middle section is being shaded, start should be the smaller value.}

\item{stop}{the number indicating where the shading stops for shading the middle of a distribution.}

\item{prob}{the area or probability to be shaded.  This value should be the area and not the percentile (i.e. shading the area to the right of the 97.5\% would take 0.025).}

\item{mu}{the mean of the distribution if it isn't a standardized distribution.}

\item{sd}{the standard deviation of the distribution if it isn't a standardized distribution.}

\item{n}{the sample size if the start/stop/prob come from (or relate to) a sample.}

\item{direction}{the area to be shaded.  direction takes 4 possible values: 1"greater", 2"less",3 "middle", 4"both".  "greater" = the right tail.  "less" = the left tail.  "middle" = the middle of the distribution.  "both" = both the left and right tails.}

\item{test}{the distribution to be used.  Takes either "z" or "t".}

\item{fill}{the color to be used for shading.}

\item{digits}{an integer in for the number of decimal places to which the answer should be rounded.  Note: positive numbers refer to the number of digits after the decimal place, negative numbers refer to the number of digits before the decimal place, and 0 refers to the nearest whole number.}
}
\value{
a plot with a shaded area.
}
\description{
Shades an area under a density curve corresponding the being tested in an introductory level 1-sample hypothesis test of means.  Either a probability value or an X value can be provided.
Either start or prob must be provided to indicate an area to be shaded.  If the area to be shaded is in the middle, both start and stop or only prob are necessary.
}
\examples{
shade_area(15, mu = 10, sd = 5, direction = "greater", test = "z")
shade_area(80, mu = 100, sd = 50, n = 40, direction = "both", test = "t")

}
\seealso{
https://sebastiansauer.github.io/shade_Normal_curve/
}
